chore(): move transaction.ts to @/types
