+++ src/app/api/beep/profile/[tokenBoundAccount]/asset-transactions/deposit/route.ts
@@ -3,7 +3,7 @@ import { Alchemy, AssetTransfersCategory, SortingOrder } from 'alchemy-sdk'
 
 import { ALCHEMY_CONFIG } from '@/constants/alchemy'
 import { TOKENS_FROM } from '@/constants/token'
-import { TransactionType } from '@/constants/transactions'
+import { TransactionType } from '@/types/transactions'
 
 const alchemy = new Alchemy(ALCHEMY_CONFIG)
